+++ lib/screens/changelog_screen.dart
@@ -3,11 +3,17 @@ import 'dart:developer';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_markdown/flutter_markdown.dart';
+import 'package:privacyidea_authenticator/utils/localization_utils.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 // TODO Make accessible from settings?
 
 class ChangelogScreen extends StatelessWidget {
+  // Without the offset the scroll bar is not shown the first time the sceen
+  // is displayed. This is a workaround for that bug and hopefully works on
+  // all devices.
+  final ScrollController _controller = ScrollController(initialScrollOffset: 2);
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -24,13 +30,19 @@ class ChangelogScreen extends StatelessWidget {
           future: DefaultAssetBundle.of(context).loadString('CHANGELOG.md'),
           builder: (context, snapshot) {
             if (snapshot.hasData) {
-              return Markdown(
-                data: snapshot.data,
-                onTapLink: (String text, String href, String title) =>
-                    _launchURL(href),
+              return Scrollbar(
+                controller: _controller,
+                isAlwaysShown: true,
+                child: Markdown(
+                  controller: _controller,
+                  data: snapshot.data,
+                  onTapLink: (String text, String href, String title) =>
+                      _launchURL(href),
+                ),
               );
             }
-            return CircularProgressIndicator();
+            return Center(
+                child: Text(Localization.of(context).somethingWentWrong));
           },
         ));
   }